+++ src/lib/packages/popover/src/PopoverTarget/index.ts
@@ -1,0 +1,5 @@
+import PopoverTarget from './PopoverTarget.svelte'
+
+export {
+	PopoverTarget
+}