+++ src/lib/packages/forms/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cluue/forms
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @cluue/icons@1.0.14
+  - @cluue/base@1.0.15
+
 ## 1.0.12
 
 ### Patch Changes